% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1c.R
\name{make_abbrev_footnote}
\alias{make_abbrev_footnote}
\title{Make footnotes of abbreviations}
\usage{
make_abbrev_footnote(abbrevs, sort = TRUE, sep = "; ", end = ".")
}
\arguments{
\item{abbrevs}{Either a named \code{list} of abbreviations, where the list
names are the abbreviations and the list items are the strings being
abbreviated, or the name of a YAML file that contains such a list.}

\item{sort}{Should the abbreviations be automatically sorted?}

\item{sep}{The seperator to use between abbreviations.}

\item{end}{Punctuation to appear at the end of the footnote.}
}
\value{
A function that takes any number of arguments, each argument being
an abbreviation to include. This function, when called, will generate the
footnote string.
}
\description{
This higher order function can be used to generate functions to produce
strings that are suitable for use as footnotes in tables (or figures) that
contain abbreviations. By convention, it is required that all abbreviations
appearing in a table or figure be spelled out in full in a footnote.
}
\details{
If an abbreviation is not found in the list, the function will
generate a warning, and produce a strange HTML code that renders as a red on
yellow "XXX" in the browser (to draw attention to the issue).
}
\examples{
abbrevs <- list(
  ALP  = "alkaline phosphatase",
  ALT  = "alanine aminotransferase",
  AST  = "aspartate aminotransferase",
  BMI  = "body mass index",
  BSA  = "body surface area")

abbrev_footnote <- make_abbrev_footnote(abbrevs)
abbrev_footnote("BMI", "ALP", "ALT")

\dontrun{
abbrev_footnote("BMI", "XXX", "YYY")  # Generates a warning
}
}
