% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1c.R
\name{one_row_per_id}
\alias{one_row_per_id}
\title{Extract one row per ID from a \code{data.frame}}
\usage{
one_row_per_id(data, idcol)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}}.}

\item{idcol}{The name or numeric index of the ID column.}
}
\value{
A \code{\link[base]{data.frame}} in which each ID appears in a
single row.  Only columns that are invariant (and hence unambiguous)
within each ID level are retained.
}
\description{
A convenience function to filter a \code{data.frame} retaining one row for
each unique value of an ID column. Only columns that are invariant (and
hence unambiguouse) within each ID level are retained.
}
\examples{
library(nlme)
one_row_per_id(Phenobarb, "Subject")
}
