% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1c.R
\name{table1}
\alias{table1}
\title{Generate an HTML table of descriptive statistics}
\usage{
table1(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional aguments, see \code{\link[table1]{table1}}.}
}
\description{
The \code{table1c} package is a light wrapper around the package
\href{https://CRAN.R-project.org/package=table1}{\code{table1}}, with some
customizations for the convenience of Certara IDD. The main interface (i.e.
this function), is retained. For details, see the function in the parent
package: \code{\link[table1]{table1}}.
}
